move some AuthMiddleware to Middleware
