+++ example/src/main/scala/example/BasicAuth.scala
@@ -1,6 +1,6 @@
 package example
 
-import zhttp.http.HttpMiddleware.basicAuth
+import zhttp.http.Middleware.basicAuth
 import zhttp.http._
 import zhttp.service.Server
 import zio.{App, ExitCode, URIO}
@@ -13,7 +13,7 @@ object BasicAuth extends App {
   }
 
   // Composing all the HttpApps together
-  val app: UHttpApp = user @@ basicAuth("admin", "admin")
+  val app: UHttpApp = user @@@ basicAuth("admin", "admin")
 
   // Run it like any simple app
   override def run(args: List[String]): URIO[zio.ZEnv, ExitCode] =
