+++ zio-http/src/main/scala/zhttp/http/Middleware.scala
@@ -2,6 +2,7 @@ package zhttp.http
 
 import io.netty.handler.codec.http.HttpHeaderNames
 import zhttp.http.CORS.DefaultCORSConfig
+import zhttp.http.Headers.BasicSchemeName
 import zio.clock.Clock
 import zio.duration.Duration
 import zio.{UIO, ZIO}
@@ -177,6 +178,33 @@ object Middleware {
    * An empty middleware that doesn't do anything
    */
   def identity: Middleware[Any, Nothing, Nothing, Any, Any, Nothing] = Middleware.Identity
+
+  /**
+   * Creates an authentication middleware that only allows authenticated requests to be passed on to the app.
+   */
+  def auth(verify: Headers => Boolean, responseHeaders: Headers = Headers.empty): HttpMiddleware[Any, Nothing] =
+    Middleware.ifThenElse[Request](req => verify(req.getHeaders))(
+      _ => Middleware.identity,
+      _ => Middleware.fromHttp(Http.status(Status.FORBIDDEN).addHeaders(responseHeaders)),
+    )
+
+  /**
+   * Creates a middleware for basic authentication
+   */
+  def basicAuth(f: Header => Boolean): HttpMiddleware[Any, Nothing] =
+    auth(
+      _.getBasicAuthorizationCredentials match {
+        case Some(header) => f(header)
+        case None         => false
+      },
+      Headers(HttpHeaderNames.WWW_AUTHENTICATE, BasicSchemeName),
+    )
+
+  /**
+   * Creates a middleware for basic authentication that checks if the credentials are same as the ones given
+   */
+  def basicAuth(u: String, p: String): HttpMiddleware[Any, Nothing] =
+    basicAuth { case (user, password) => (user == u) && (password == p) }
 
   /**
    * Creates a middleware for Cross-Origin Resource Sharing (CORS).
